requirement.txt
playwright
pandas
numpy
google-auth
google-auth-oauthlib
google-api-python-client
